+++ dojo/finding/__init__.py
@@ -1,0 +1,1 @@
+import dojo.finding.admin  # noqa: F401
